后端代码-最终版/requirements.txt
acsclient==1.0.10
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cors==1.0.1
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
decorator==5.1.1
editdistance==0.8.1
exceptiongroup==1.2.2
fastapi==0.112.4
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
funasr==1.1.6
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gradio==4.43.0
gradio_client==1.3.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface==0.0.1
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.0
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
modelscope==1.18.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
netifaces==0.10.6
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
orjson==3.10.7
oss2==2.19.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.28.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.1.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.3.0
pytorch-wpe==0.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rich==13.8.0
ruff==0.6.4
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
socketio==0.2.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.32
starlette==0.38.4
sympy==1.13.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.1
torch-complex==0.4.4
torchaudio==2.4.1
tqdm==4.66.5
transformers==4.44.2
twilio==9.2.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.1
zope.event==5.0
zope.interface==7.0.3
